Mute OSError, if directory already exists.
It could happen, that two tasks of the same class, which run in parallel
and which need to create the same directory, would interfere.

One task can create the directory between the other tasks' calls
to `os.path.exists` and `os.makedirs`.

Workaround this issue by muting the `OSError`, if `errno == 17` (errno.EEXIST).